+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+## [0.27.1] - 2025-12-08
+
 ### Added
 - `pw.Table.filter_out_results_of_forgetting` method, allowing to revert the effects of forgetting at a later stage.